Changes to be committed: 	modified:   templates/biblioteca/admin/edit_item.html 	modified:   templates/biblioteca/admin/view_items.html 	modified:   views.py
The view_items page now has a default 'splashdown' display that prompts the user to select a type of item to view them.

The edit_item page reroutes to the default splashdown page.

TODO: give the back-buttons a form submit with the appropriate data to bring them to the desired page