Changes to be committed: 	modified:   templates/biblioteca/admin/add_item.html 	modified:   templates/biblioteca/admin/edit_item.html 	new file:   templates/biblioteca/admin/item_back_button.html
This commit solves Issue #48. It adds a back-button template that should work for any page that shows items.

It updates the back button from a hyperlink to accessing the appropriate views.py function to render the page. The previous page from add_item or edit_item page is then displayed.