Changes to be committed: 	modified:   templates/biblioteca/client/view_items.html 	new file:   templates/biblioteca/item_base.html
 Changes not staged for commit:
	deleted:    templates/biblioteca/item_base.html

There is no complex back button behaviour currently required for the clients. Will likely revisit in a future sprint.

item_base.html is a parent template for add_item.html and edit_item.html, who both inherit the button from the template.

there is a default state that allows the admin to select their item_type as desired. The item_type of the admin is tracked and moves the admin to appropriate previous page